pyproject.toml
[tool.poetry]
name = "OAtlas"
version = "0.1.0"
description = ""
authors = ["pUrGe12 <achintya.jai@owasp.org>"]
keywords = [
    "osint",
    "reconnaissance",
    "information-gathering",
    "security",
    "open-source-intelligence",
    "atlas"
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = ">=3.10"
requests = ">=2.32.4,<3.0.0"
google-cloud-aiplatform = ">=1.105.0,<2.0.0"
dnspython = ">=2.7.0,<3.0.0"
pre-commit = ">=4.2.0,<5.0.0"
deepface = { git = "https://github.com/serengil/deepface.git", rev = "master" }
protobuf = "^3.20"
numpy = "<2"
pillow = "^11.3.0"
sqlalchemy = "^2.0.42"
apsw = "^3.50.4.0"
maturin = "^1.9.2"
torch = "^2.7.1"
playwright = "^1.54.0"
pydantic = "^2.11.7"
python-magic = "^0.4.27"
pytesseract = "^0.3.13"
streamlit = "^1.48.1"
python-dotenv = "^1.1.1"
aiohttp = "^3.12.15"
asyncio = "^4.0.0"
ipaddr = "^2.2.0"
multiprocess = "^0.70.18"
netaddr = "^1.3.0"
paramiko = "^4.0.0"
py3dns = "^4.0.2"
pyopenssl = "^25.3.0"
pysocks = "^1.7.1"
pyyaml = "^6.0.2"
uvloop = "^0.21.0"
zipp = "^3.23.0"
c2pa-python = "^0.23.1"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.2.1,<0.8.0"
isort = "^5.13.2"
pre-commit = "^4.2.0"
ipython = "^8.16.1"
maturin = "^1.9.2"

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = ["oatlas"]
no_inline_sort = true
skip = ["__init__.py"]

[tool.ruff]
line-length = 99
per-file-ignores = { "__init__.py" = ["ALL"], "oatlas.py" = ["E402"], "oatlas/tools/AI_image_detector/AI_image_analysis.py" = ["E402"], "oatlas/tools/image_analysis/deep_learning_scans.py" = ["E402"]}

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.maturin]
manifest-path = "oatlas/rust/Cargo.toml"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
